Fix bugs and auto-indent
