+++ src/models/userModel.ts
@@ -3,10 +3,9 @@ import { Schema, model } from 'mongoose';
 const userSchema = new Schema({
   name: { type: String, required: true },
   email: { type: String, required: true, unique: true },
-  password: { type: String, required: true },
-  savedSearches: { type: Array, required: false },
+  password: { type: String, required: true }
 });
 
-const UserModel = model('DRI_Users', userSchema);
+const UserModel = model(process.env.USERS_COLLECTION, userSchema);
 
 export default UserModel;
