+++ src/routes/geolocationRoute.ts
@@ -1,8 +1,8 @@
 import { Router } from "express";
-import * as geolocalisationControllers from "../controllers/geolocalisationControllers";
-import { authMiddleware } from "../middlewares/authMiddlewares";
+import * as geolocalisationControllers from "../controllers/geolocationController";
+import { authMiddleware } from "../middlewares/authMiddleware";
 
-export function getRouter(){
+export function getRouter() {
     const router = Router();
 
     router.get('/geolocalisation/:dpe/:ges/:zipcode/:surface', authMiddleware, geolocalisationControllers.getGeolocalisation);
