+++ src/utils/authUtils.ts
@@ -17,9 +17,9 @@ export function createTokens(user: any) {
 
 export async function hashPassword(password: string) {
     const salt = await bcrypt.genSalt(10);
-    
+
     const hashedPassword = await bcrypt.hash(password, salt);
-    
+
     return hashedPassword;
 }
